requirements.txt
# Core Framework
streamlit>=1.31.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.20
langchain-groq>=0.0.1

# LLM & Embeddings
groq>=0.4.0
sentence-transformers>=2.2.0

# Vector Store
chromadb>=0.4.0

# Document Processing
pypdf>=3.17.0

# Data Generation
faker>=22.0.0

# Environment
python-dotenv>=1.0.0

# Utilities
pydantic>=2.0.0
